feat: New delete

+++ server.ts
@@ -62,9 +62,19 @@ app.delete("/api/holiday-plans/:id", async (req: Request, res: Response) => {
   const { id } = req.params;
 
   try {
+    const holidayPlan = await prisma.holidayPlan.findUnique({
+      where: { id },
+    });
+
+    if (!holidayPlan) {
+      return res.status(404).send("Plano de férias não encontrado.");
+    }
+
     await prisma.holidayPlan.delete({
       where: { id },
     });
+
+    console.log(`Plano de férias com ID ${id} excluído com sucesso.`); // Adicionando um log para registrar a exclusão bem-sucedida
     res.status(204).send();
   } catch (err) {
     console.error("Erro ao excluir o plano de férias: ", err);
